Completed fans project

+++ Programmazione 2/IntelliJ/SoloVentilatori/src/main/java/model/interfaces/PowerInterface.java
@@ -2,5 +2,5 @@ package model.interfaces;
 
 public interface PowerInterface {
     String getPower();
-    double getMultiplier();
+    int getMultiplier();
 }
